Add files via upload
The dice game starts by asking the user how many players are going to play.  After that it gives the opportunity to each player to throw the dice and records each-player´s result in a list.  When all the players have taken their opportunity it compares the results and says the winner.  In case there´s a tie, it shows which participants tied, and they play an extra round, recalling them (I used a recurring function).